servo: Include fractional nanoseconds within clock offset
Pass the clock offset to servo_sample() as a floating point value,
allowing the servo to take fractional nanoseconds into account.

Note that a floating point value can sensibly hold a clock offset, but
cannot sensibly hold an absolute time due to its limited range.
Servos that perform calculations on absolute times therefore cannot
make use of fractional nanoseconds.  At present, only the PI servo is
structured in a way that allows fractional nanoseconds to be used.

+++ servo_private.h
@@ -30,7 +30,7 @@ struct servo {
 	void (*destroy)(struct servo *servo);
 
 	double (*sample)(struct servo *servo,
-			 int64_t offset, uint64_t local_ts, double weight,
+			 double offset, uint64_t local_ts, double weight,
 			 enum servo_state *state);
 
 	void (*sync_interval)(struct servo *servo, double interval);
